Set sprockets vertical object created
