+++ UIElements/calibrateLengthsPopup.py
@@ -7,40 +7,26 @@ from   kivy.uix.gridlayout                       import   GridLayout
 from   kivy.properties                           import   ObjectProperty
 from   kivy.properties                           import   StringProperty
 from   kivy.uix.popup                            import   Popup
+from   UIElements.setSprocketsVertical           import   SetSprocketsVertical
 
 class CalibrateLengthsPopup(GridLayout):
     done   = ObjectProperty(None)
     
-    def LeftCW(self):
-        self.data.gcode_queue.put("G91 ")
-        self.data.gcode_queue.put("B09 L.5 ")
-        self.data.gcode_queue.put("G90 ")
     
-    def LeftCCW(self):
-        self.data.gcode_queue.put("G91 ")
-        self.data.gcode_queue.put("B09 L-.5 ")
-        self.data.gcode_queue.put("G90 ")
+    def establishDataConnection(self, data):
+        '''
         
-    def RightCW(self):
-        print "right CW"
-        self.data.gcode_queue.put("G91 ")
-        self.data.gcode_queue.put("B09 R-.5 ")
-        self.data.gcode_queue.put("G90 ")
-    
-    def RightCCW(self):
-        self.data.gcode_queue.put("G91 ")
-        self.data.gcode_queue.put("B09 R.5 ")
-        self.data.gcode_queue.put("G90 ")
+        Sets up the data connection between this popup and the shared data object
+        
+        '''
+        self.data = data
+        self.setSprocketsVertical.data = data
+        self.setSprocketsVertical.carousel = self.carousel
     
     def stop(self):
         self.data.quick_queue.put("!") 
         with self.data.gcode_queue.mutex:
             self.data.gcode_queue.queue.clear()
-    
-    def setZero(self):
-        #mark that the sprockets are straight up
-        self.data.gcode_queue.put("B06 L0 R0 ");
-        self.carousel.load_next()
     
     def calibrateChainLengths(self):
         self.data.gcode_queue.put("B02 ")
