+++ UIElements/diagnosticsMenu.py
@@ -49,7 +49,9 @@ class Diagnostics(FloatLayout, MakesmithInitFuncs):
         self.data.gcode_queue.put("B06 L0 R0 ");
         
         self.popupContent      = CalibrateLengthsPopup(done=self.dismissMeasureMachinePopup)
-        self.popupContent.data = self.data
+        
+        self.popupContent.establishDataConnection(self.data)
+        
         self._popup = Popup(title="Calibrate Chain Lengths", content=self.popupContent,
                             size_hint=(0.85, 0.95), auto_dismiss = False)
         self._popup.open()
